[common] Updated to use Profiler parser
To parse the log profiler output, we use some code from the Profiler project. This allow use to write less code and keep it more up to date with the log profiler protocol. For now, there is no shared repository, so it will have to be updated by copy pasting the desired version.